requirements.txt
numpy
pandas
matplotlib
tensorflow==2.16.1
scikit-learn
keras==2.15.0
pillow
tqdm
fastapi
uvicorn
python-multipart
pytest
httpx